requirements.txt
twscrape>=0.10.0
httpx>=0.25.0
python-dotenv>=1.0.0
tweepy>=4.14.0
schedule>=1.2.0 
google-genai